Check kube-proxy health on linux
